% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_operator.R
\name{f_operator}
\alias{f_operator}
\title{Function Operator}
\usage{
f_operator(f)
}
\arguments{
\item{f}{A function}
}
\value{
A function
}
\description{
If input is greater than 0, then return normal output. If input is invalid, return an error condition object, with “invalid_input” subclass and invalid value attached.
}
\examples{
sqrt.3 = f_operator(sqrt(-1))
}
