% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_2.R
\name{log.2}
\alias{log.2}
\title{log.2}
\usage{
\method{log}{`2`}(x)
}
\arguments{
\item{x}{A number}
}
\value{
An error or a number
}
\description{
A way to return errors for log of negative numbers instead of NAs
}
\examples{
\donttest{log.2(2)
}
\donttest{log.2(-1)
}

}
