% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt_2.R
\name{sqrt.2}
\alias{sqrt.2}
\title{sqrt.2}
\usage{
\method{sqrt}{`2`}(x)
}
\arguments{
\item{x}{A number}
}
\value{
An error or a number
}
\description{
A way to return errors for square root of negative numbers instead of NAs
}
\examples{
sqrt.2(4)
}
